Update compounds.yml
Please do not make live pending more verification. However, I believe the dGH calculations were incorrect. The way to calculate dGH is...

CaCO3 = 2.5 * Ca + 4.10 * Mg + 2.18 * Na + 1.28 * K + 1.79 * Fe + 1.82 * Mn

This is when the things that make up GH are known which in this case they are.

I updated the dGH value and/or added dGH if it contains the above elements.